add userendpoint

+++ src/main/java/dateregistration/DateRegistrationEndpoint.java
@@ -2,6 +2,7 @@ package dateregistration;
 
 //import com.sun.org.glassfish.gmbal.ParameterNames;
 
+import login.User;
 import org.junit.runners.Parameterized;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 
 import javax.validation.Valid;
+import java.security.Principal;
 import java.util.List;
 
 /**
@@ -36,17 +38,22 @@ public class DateRegistrationEndpoint {
 
     //set dates
     @RequestMapping(method = RequestMethod.POST, value = {"/user/{userID}/dateregistration"})
-    public ResponseEntity<Void> postCheckinDate(@RequestBody @Valid DateRegistration dateRegistration, @PathVariable("userID") int userID) throws InvalidDateException {
-        if (dateRegistration.getCheckinDate().isBefore(dateRegistration.getCheckoutDate())) {
-            System.out.println("=========Checkin and Checkout date==========");
-            System.out.println("Checkin Date: " + dateRegistration.getCheckinDate());
-            System.out.println("Checkout Date: " + dateRegistration.getCheckoutDate());
-            database.insertData(dateRegistration,userID);
-            return new ResponseEntity<>(HttpStatus.CREATED);
-        } else {
-          //throw new InvalidDateException();
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        }
+    public ResponseEntity<Void> postCheckinDate(@RequestBody @Valid DateRegistration dateRegistration, @PathVariable("userID") int userID
+    , Principal principal) throws InvalidDateException {
+        User currentUser=database.loadUserByUsername(principal.getName());
+        if (currentUser.getUserId()==userID ) {
+            if (dateRegistration.getCheckinDate().isBefore(dateRegistration.getCheckoutDate())) {
+                System.out.println("=========Checkin and Checkout date==========");
+                System.out.println("Checkin Date: " + dateRegistration.getCheckinDate());
+                System.out.println("Checkout Date: " + dateRegistration.getCheckoutDate());
+                database.insertData(dateRegistration, userID);
+                return new ResponseEntity<>(HttpStatus.CREATED);
+            } else {
+                //throw new InvalidDateException();
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+        }else
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
 
@@ -59,17 +66,25 @@ public class DateRegistrationEndpoint {
 
     //get booking overview from one user
     @RequestMapping(method = RequestMethod.GET, value = {"user/{userID}/dateregistration"})
-    public ResponseEntity<List<DateRegistration>> listAllBooking(@PathVariable ("userID") int userID) {
-        System.out.println("Booking Overview " + database.showBookingDates(userID));
-        return new ResponseEntity<>(database.showBookingDates(userID), HttpStatus.OK);
+    public ResponseEntity<List<DateRegistration>> listAllBooking(@PathVariable ("userID") int userID, Principal principal) {
+        User currentUser= database.loadUserByUsername(principal.getName());
+        if(currentUser.getUserId()==userID) {
+            System.out.println("Booking Overview " + database.showBookingDates(userID));
+            return new ResponseEntity<>(database.showBookingDates(userID), HttpStatus.OK);
+        }else
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
-    //delete all booking
-    @RequestMapping(method = RequestMethod.DELETE, value = {"/dateregistration"})
-    public ResponseEntity<Void> deleteAll() {
+    //delete all booking from one user
+    @RequestMapping(method = RequestMethod.DELETE, value = {"/user/{userID}/dateregistration"})
+    public ResponseEntity<Void> deleteAll(@PathVariable ("userID") int userID, Principal principal) {
+        User currentUser= database.loadUserByUsername(principal.getName());
+        if (currentUser.getUserId()==userID){
         database.deleteData();
         System.out.println("All bookings are deleted.");
         return new ResponseEntity<>(HttpStatus.OK);
+        }else
+            return new ResponseEntity<Void>(HttpStatus.UNAUTHORIZED);
     }
 
     //delete one booking
